chore: move shell expansion to `Config::validate()`
